+++ src/standard/queue.rs
@@ -1,7 +1,7 @@
 //! An OpenCL command queue.
 
 use std;
-use core::{self, CommandQueue as CommandQueueCore, DeviceId as DeviceIdCore, Context as ContextCore,
+use core::{self, CommandQueue as CommandQueueCore, Context as ContextCore,
     CommandQueueInfo, CommandQueueInfoResult};
 use standard::{self, Context, Device};
 
@@ -16,7 +16,7 @@ use standard::{self, Context, Device};
 pub struct Queue {
     obj_core: CommandQueueCore,
     context_obj_core: ContextCore,
-    device_id_core: DeviceIdCore,
+    device: Device,
 }
 
 impl Queue {
@@ -36,18 +36,18 @@ impl Queue {
         // assert!(device_ids_core.len() == 1, "Queue::new_by_device_index: Error resolving device ids.");
         // let device_id_core = device_ids_core[0].clone();
 
-        let device_id_core = match device {
-            Some(d) => d.as_core().clone(),
-            None => context.get_device_by_index(0).as_core().clone(),
+        let device = match device {
+            Some(d) => d.clone(),
+            None => context.get_device_by_index(0).clone(),
         };
 
-        let obj_core = core::create_command_queue(context.core_as_ref(), &device_id_core)
+        let obj_core = core::create_command_queue(context.core_as_ref(), &device)
             .expect("[FIXME: TEMPORARY]: Queue::new_by_device_index():"); 
 
         Queue {
             obj_core: obj_core,
             context_obj_core: context.core_as_ref().clone(),
-            device_id_core: device_id_core, 
+            device: device, 
         }
     }
 
@@ -67,17 +67,17 @@ impl Queue {
             None => Vec::with_capacity(0),
         };
 
-        let device_ids_core = context.resolve_device_idxs(&device_idxs);
-        assert!(device_ids_core.len() == 1, "Queue::new_by_device_index: Error resolving device ids.");
-        let device_id_core = device_ids_core[0].clone();
+        let devices = context.resolve_device_idxs(&device_idxs);
+        assert!(devices.len() == 1, "Queue::new_by_device_index: Error resolving device ids.");
+        let device = devices[0].clone();
 
-        let obj_core = core::create_command_queue(context.core_as_ref(), &device_id_core)
+        let obj_core = core::create_command_queue(context.core_as_ref(), &device)
             .expect("[FIXME: TEMPORARY]: Queue::new_by_device_index():"); 
 
         Queue {
             obj_core: obj_core,
             context_obj_core: context.core_as_ref().clone(),
-            device_id_core: device_id_core, 
+            device: device, 
         }
     }  
 
@@ -101,8 +101,8 @@ impl Queue {
     ///
     /// Not to be confused with the zero-indexed `device_idx` passed to `::new()`
     /// when creating this queue.
-    pub fn device_core_as_ref(&self) -> &DeviceIdCore {
-        &self.device_id_core
+    pub fn device(&self) -> &Device {
+        &self.device
     }
 
     /// Returns info about this queue.